+++ client-modern/src/main/java/com/paraguacraft/pvp/modern/ParaguacraftPvPModern.java
@@ -24,7 +24,7 @@ import net.minecraft.client.MinecraftClient;
 public class ParaguacraftPvPModern implements ClientModInitializer {
 
     public static final String MOD_ID = "paraguacraftpvp-modern";
-    public static final String VERSION = "0.9.31";
+    public static final String VERSION = "0.9.32";
 
     @Override
     public void onInitializeClient() {