+++ client/src/main/java/com/paraguacraft/pvp/ParaguacraftPvP.java
@@ -37,7 +37,7 @@ public class ParaguacraftPvP {
 
     public static final String MODID = "paraguacraftpvp";
     public static final String NAME = "Paraguacraft PvP Client";
-    public static final String VERSION = "2.1.57";
+    public static final String VERSION = "2.1.58";
 
     @Mod.Instance(MODID)
     public static ParaguacraftPvP instance;